MoviesPresenter creating domain package and moving model class to it
