+++ app/src/main/java/com/tomerpacific/moviepresenter/model/MainUiState.kt
@@ -1,5 +1,7 @@
 package com.tomerpacific.moviepresenter.model
 
+import com.tomerpacific.moviepresenter.domain.MovieModel
+
 data class MainUiState(
     val isLoading: Boolean = true,
     val moviesList: List<MovieModel>? = null,
