+++ app/src/main/java/com/tomerpacific/moviepresenter/repository/MovieRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.tomerpacific.moviepresenter.repository
 
 import android.graphics.BitmapFactory
 import com.tomerpacific.moviepresenter.BuildConfig
-import com.tomerpacific.moviepresenter.model.MovieModel
+import com.tomerpacific.moviepresenter.domain.MovieModel
 import com.tomerpacific.moviepresenter.model.TMDBResponse
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.coroutineScope
